been a while
I've waited a long time because of a bug in angelica. The bug is still not fixed but I got tired of waiting: if you log into a world and get a black screen, then the solution is use windowed mode and slightly resize manually the window. This causes a screen redraw that fixes the black screen... Until the next boot.

Of course this has months of mod updates, but should be fine.